+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Change log
 v 0.3.x
 -------
 
+* Creates `Synfit` (`s4.fitting.Synfit`), a routine to automatically fit
+   spectral lines or wavelength regions.
 * Merge `linedid_plot <https://github.com/phn/lineid_plot>`_
   (`PR #23 <https://github.com/gabraganca/S4/pull/23>`_).
 * Bind Travis CI partially (see `PR #24 <https://github.com/gabraganca/S4/pull/24>`_